Rename api to routes, convert webhook handlers to routers
Move eggnet, samsara, and subs into route files that combine the
business logic with the route definition. Merge public and admin
routes for raffle and verified into single files with two exports.
app.ts is now purely middleware setup and route mounting.
