--- packages/oaf/src/server/apps/oaf/api/raffle.ts
@@ -1,10 +1,0 @@
-import { Router } from "express";
-
-import { getRafflesWithWinners } from "../../../../clients/database.js";
-
-export const raffleRouter = Router();
-
-raffleRouter.get("/", async (_req, res) => {
-  const raffles = await getRafflesWithWinners();
-  res.json({ raffles });
-});
