--- packages/oaf/src/server/apps/oaf/api/verified.ts
@@ -1,10 +1,0 @@
-import { Router } from "express";
-
-import { getVerifiedPlayers } from "../../../../clients/database.js";
-
-export const verifiedRouter = Router();
-
-verifiedRouter.get("/", async (_req, res) => {
-  const players = await getVerifiedPlayers();
-  res.json({ players });
-});
